splits ifElse into ifThen and ifThenElse
closes #104
